requirements.txt
asgiref==3.11.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
Django==5.2.13
django-allauth==65.16.1
django-allauth-ui==1.8.1
django-widget-tweaks==1.5.1
djangorestframework==3.15.0
idna==3.13
iniconfig==2.3.0
packaging==26.1
pluggy==1.6.0
pycparser==3.0
Pygments==2.20.0
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.33.1
responses==0.25.8
slippers==0.6.3
sqlparse==0.5.5
typeguard==2.13.3
typing_extensions==4.15.0
urllib3==2.6.3
